+++ project/src/const.ts
@@ -62,3 +62,12 @@ export enum APIRoute {
   Favorites = '/favorites',
   Comments = '/comments',
 }
+
+export const errorMessages = {
+  authorization: 'Вы не авторизованы, не забудьте авторизоваться.',
+  login: 'Не получилось авторизоваться, попробуйте еще раз.',
+  logout: 'Не удается выйти из аккаунта, попробуйте еще раз.',
+  comments: 'Не удалось загрузить отзывы, попробуйте перезагрузить страницу.',
+  nearbyOffers: 'Не удалось загрузить места поблизости, попробуйте перезагрузить страницу.',
+  postComment: 'Не удалось загрузить отзыв, попробуйте еще раз',
+};
