+++ project/src/index.tsx
@@ -6,8 +6,9 @@ import { rootReducer } from './store/root-reduser';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import { createAPI } from './services/api';
-import { requireAuthorization } from './store/action';
-import { fetchOffersAction, checkAuthAction } from './store/api-action';
+import { requireAuthorization } from './store/user-process/action';
+import { fetchOffersAction } from './store/offers-data/api-action';
+import { checkAuthAction } from './store/user-process/api-action';
 import { AuthorizationStatus } from './const';
 import { offers } from './mocks/offers';
 import { redirect } from './store/middlewares/redirect';
