+++ project/src/pages/login/login.tsx
@@ -1,14 +1,9 @@
 import Header from '../../components/header/header';
 import LoginForm from '../../components/login-form/login-form';
 import { Link } from 'react-router-dom';
-import { AppRoute, CitiesList } from '../../const';
+import { AppRoute } from '../../const';
 import { citiesList } from '../../const';
-
-
-function getRandomCity ( obj: CitiesList ): string  {
-  const keys = Object.keys(obj);
-  return keys[Math.floor(Math.random()*keys.length)];
-}
+import { getRandomCity } from '../../util';
 
 function Login(): JSX.Element {
   return (
