+++ project/src/store/ui-state/action.ts
@@ -1,0 +1,16 @@
+import { createAction } from '@reduxjs/toolkit';
+import { ActionType } from '../../types/action';
+
+export const cityChange = createAction(
+  ActionType.СityСhange,
+  (city: string) => ({
+    payload: city,
+  }),
+);
+
+export const sortTypeChange = createAction(
+  ActionType.SortTypeChange,
+  (sortType: string) => ({
+    payload: sortType,
+  }),
+);
